token-verifier/requirements.txt
boto3==1.34.135
cryptography==42.0.8
PyJWT==2.3.0
